--- receiver/telemetrygeneratorreceiver/generator.go
@@ -1,9 +1,0 @@
-package telemetrygeneratorreceiver
-
-// type generator interface {
-// 	// Passes in a component.Type which has a value of logs, metrics, traces. Returns true or false if this generator supports it.
-// 	SupportsType(component.Type) bool
-// 	GenerateMetrics() pmetric.Metrics
-// 	GenerateLogs() plog.Logs
-// 	GenerateTraces() ptrace.Traces
-// }
